Kaggle Bike Share Demand Challenge
This is a Kaggle Challenge project. Kaggle provided a bicycle rental dataset for a large city. The challenge is to build a machine learning model that can predict the number of bikes to be rented. Various weather conditions and bike rental counts are given for the first 20 days of each month for two years. The challenge is to predict the remaining 10 days of the month for each month.